Update fluentd.service.j2
Delete ExecStop as it can trigger an error depending on the OS layout. kill location isn't a reliable thing in the first place (it's living at /bin on Ubuntu), but systemd is able to kill processes without being told how anyway. KillSignal should be used if a SIGKILL signal is really necessary.